[Refactor(BookSearchRepositoryTest)] 테스트 코드 복구
