+++ core/data/booksearch/internal/src/androidTest/java/com/kova700/amazonbookstorepractice/core/data/booksearch/internal/ExampleInstrumentedTest.kt
@@ -1,12 +1,10 @@
-package com.kova700.booksearch
+package com.kova700.amazonbookstorepractice.core.data.booksearch.internal
 
-import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
-
+import androidx.test.platform.app.InstrumentationRegistry
+import org.junit.Assert.assertEquals
 import org.junit.Test
 import org.junit.runner.RunWith
-
-import org.junit.Assert.*
 
 /**
  * Instrumented test, which will execute on an Android device.
@@ -19,6 +17,9 @@ class ExampleInstrumentedTest {
 	fun useAppContext() {
 		// Context of the app under test.
 		val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-		assertEquals("com.kova700.booksearch.test", appContext.packageName)
+		assertEquals(
+			"com.kova700.amazonbookstorepractice.core.data.booksearch.internal.test",
+			appContext.packageName
+		)
 	}
 }
