+++ core/network/booksearch/src/main/java/com/kova700/core/network/booksearch/BookSearchNetworkService.kt
@@ -1,12 +1,13 @@
 package com.kova700.core.network.booksearch
 
+import com.kova700.amazonbookstorepractice.core.data.booksearch.external.model.KakaoBookSearchSortType
 import com.kova700.core.network.booksearch.model.BookResponse
 
 interface BookSearchNetworkService {
 
 	suspend fun searchBooks(
 		query: String,
-		sort: String,
+		sort: KakaoBookSearchSortType,
 		page: Int,
 		size: Int,
 	): BookResponse
