+++ gradle/libs.versions.toml
@@ -12,7 +12,7 @@ androidGradlePlugin = "8.5.2"
 espressoCore = "3.6.1"
 junit = "4.13.2"
 kotlin = "2.0.10"
-hilt = "2.49"
+hilt = "2.51.1"
 hiltNavigationCompose = "1.2.0"
 ksp = "2.0.10-1.0.24"
 androidxCore = "1.13.1"
@@ -21,7 +21,7 @@ androidxActivity = "1.9.1"
 compose = "1.6.8"
 composeMaterial3 = "1.2.1"
 navigationCompose = "2.7.7"
-coilCompose = "2.5.0"
+coilCompose = "2.6.0"
 kotlinxCollectionsImmutable = "0.3.7"
 kotlinxSerializationJson = "1.6.3"
 
@@ -29,14 +29,14 @@ ktor = "2.3.7"
 
 datastorePreferences = "1.1.1"
 
-mockito = "5.7.0"
+mockito = "5.2.0"
 mockitoKotlin = "5.2.1"
 testAndroidxCore = "1.6.1"
 androidxTestJunit = "1.2.1"
-coroutines = "1.7.3"
+coroutines = "1.8.0"
 appcompat = "1.7.0"
 material = "1.12.0"
-org-jetbrains-kotlin-jvm = "1.9.0"
+org-jetbrains-kotlin-jvm = "2.0.0"
 kotlinGradlePlugin = "1.9.0"
 kotlinGradlePluginVersion = "1.9.0"
